requirements.txt
fastapi[standard]== 0.115.6
pydantic>=2.7.0,<3.0.0
SQLAlchemy~=2.0.36
pyjwt==2.10.1
passlib~=1.7.4